+++ cirq/protocols/mul_protocol.py
