+++ cirq/protocols/mul_protocol_test.py
