+++ cirq/protocols/pow_protocol.py
