+++ cirq/protocols/pow_protocol_test.py
